Move out DragService into its own file
